requirements.txt
fastapi==0.110.0
gensim==4.3.2
loguru==0.7.2
nltk==3.8.1
numpy==1.26.0
pandas==2.2.1
Pillow==10.2.0
pytest==8.1.1
Requests==2.31.0
scikit_learn
ultralytics==8.1.23
Unidecode==1.3.7
uvicorn==0.28.0
python-multipart